+++ src/views/Idea/IdeaView.tsx
@@ -1,7 +1,6 @@
 import { IdeaForms } from '@/components/DataForms';
-import { ApprovalCard, IdeaBubble, VotingCard, VotingResults } from '@/components/Idea';
+import { ApprovalCard, IdeaBubble } from '@/components/Idea';
 import IdeaBubbleSkeleton from '@/components/Idea/IdeaBubble/IdeaBubbleSkeleton';
-import VotingQuorum from '@/components/Idea/VotingQuorum';
 import { deleteIdea, getIdea, getIdeaBoxes } from '@/services/ideas';
 import { getRoom } from '@/services/rooms';
 import { getQuorum } from '@/services/vote';
@@ -113,23 +112,15 @@ const IdeaView = () => {
 
   return !isLoading && idea ? (
     <Stack width="100%" height="100%" overflow="auto" gap={2}>
-      {phase === '30' && idea.approved > 0 && <VotingCard onReload={fetchIdea} />}
-      {phase === '40' && <VotingResults idea={idea} onReload={fetchIdea} quorum={quorum} />}
       <IdeaBubble
         idea={idea}
+        quorum={quorum}
+        phase={Number(phase) as RoomPhases}
         onEdit={() => setEdit(idea)}
         onDelete={() => ideaDelete(idea.hash_id)}
+        onReload={onClose}
         disabled={Number(phase) >= 20}
-      >
-        {idea.approved >= 0 && (
-          <VotingQuorum
-            quorum={quorum}
-            phase={Number(phase) as RoomPhases}
-            votes={Number(phase) >= 30 ? Number(idea.number_of_votes) : Number(idea.sum_likes)}
-            users={Number(idea.number_of_users)}
-          />
-        )}
-      </IdeaBubble>
+      ></IdeaBubble>
       {Number(phase) >= 20 && <ApprovalCard idea={idea} onReload={fetchIdea} />}
       <Stack px={2} flex={1}>
         <CommentView />